roachtest: rename upgrade -> autoupgrade
The "upgrade" roachtest is renamed to "autoupgrade" and moved into a new
file. The reason is that this test isn't particularly well-maintained or
extensible, and so we want to "hide" it and favor the mixed-version test
harness left in upgrade.go more prominently.

Release note: None
