+++ pkg/cmd/roachtest/registry.go
@@ -18,6 +18,7 @@ func registerTests(r *testRegistry) {
 	registerAcceptance(r)
 	registerAllocator(r)
 	registerAlterPK(r)
+	registerAutoUpgrade(r)
 	registerBackup(r)
 	registerCancel(r)
 	registerCDC(r)
@@ -85,7 +86,6 @@ func registerTests(r *testRegistry) {
 	registerKVBench(r)
 	registerTypeORM(r)
 	registerLoadSplits(r)
-	registerUpgrade(r)
 	registerVersion(r)
 	registerYCSB(r)
 	registerTPCHBench(r)
